requirements.txt
openai
langchain
langchain_openai
langchain_core
langchain_community
langchainhub
pinecone-client
python-dotenv
tiktoken
docx2txt
pypdf
numpy
chromadb
streamlit
langserve
fastapi
uvicorn
sse_starlette
pysqlite3-binary